Added some rudimentary statistical analysis.
There is a new Python file, hrse.py, which contains a function that counts the
number of runs in the sequence and the length of the longest (runs()).  The
basis for the use of this test, along with a basic count of the number of 0s
and 1s, as rudimentary randomness tests was taken from a website that I
discovered tonight by Dr. Christopher Wetzel, found here:
http://faculty.rhodes.edu/wetzel/random/mainbody.html.

I also fixed a bug that was preventing the "Done" button from working on
mobile.