Added participant ID recognition on front page
The participant (née "subject") ID is now written to the page via an
AJAX call to the database.  If the browser fingerprint doesn't already
exist in the participant table, a new participant is automatically created.

Also added the scss/app.scss file to the tree.  Should've been there already,
but it is at this point simply a duplicate of the one that was already in
the prototype/ directory.