Added new footer bar, cleaned up formatting some, enhanced database.
I modified the schema to include a submission timestamp and user-agent string
for all sequence submissions.

I changed the text for step 1 in the instructions to be shorter for smaller
devices (< 640px screen width).

I added a more refined bottom navigation bar with new images, although the
images are not clickable yet.