Fixed the myresults() click issue added a bunch of button graphics and progressed further toward a myresults page.
The myresults() issue was that I had named the img id field the same as the
onclick method.  Javascript gets confused when you do that.

The templates/myresults.html file is just there purely as a debugging page for
now.  I still currently plan on generating that template on the fly.