Added new functions to support autogeneration of charts and finished the demo form.
The demographic form now has names for all of the inputs and the values are
submitted as a JSON document, along with the sequence and fingerprint, which
are carried over from the index page.

I also added a submit button that works like the done button on the index
page, using media queries.

The next step is to submit the form information into the database and generate
some rudimentary charts for the results page.