Minor cosmetic changes and bug fixes.
Replaced the white 0 buttons with ones that didn't include text indicating
what size they are.

Fixed a JS bug that was preventing the 'about' and 'overall' buttons from
working.

Log message changes in hrse.py and query.py.

Fixed a bug that prevented the 'new sequence' button from working when the
randomhuman.org name is not assigned to the instance.